iterative multi pass disassembler
This PR introduces a new framework for project reconstruction,
that will take multiple passes to recover the project.

Previously we had only one pass, with a fixed set of inputs, that cannot
have circular references. Now the inputs can be derived from the
outputs. The reconstruction will continue until a fixpoint is reached.

Each source of information, is represented by a stream.  They can be
merged in arbitrary ways using stream combinators. The details follow.

Event system
------------

A new event system is introduced by this PR. A stream of extensible
events can be fed by any program component.

Logging infrastructure
----------------------

This PR adds a logging infrastructure based on top of the event system.
The `Self()` generative functor will bring into scope 4 new functions:

 - debug
 - info
 - warning
 - error

The `debug` function would be no-op if logging is disable for the
component. The logging can be controlled with environment variables.
To enable logging of all components define `BAP_DEBUG` variable. To
enable logging of a specific component with name <name>, define
`BAP_DEBUG_<NAME>` variable, e.g., to debug `ida` plugin, set:
`BAP_DEBUG_IDA=1`, or just `BAP_DEBUG_IDA=`.

As it was said, the logging is built on top of the event stream. The
stream is processed by a `Bap_log` module, that prints all log events
into a log file, that is located by default at `~/.local/state/bap/log`
folder. Every new start of a program, creates a new log. Previous logs
are rotated. The `Bap_log` module will also print `error` events into
the `stderr`.

IDA plugin policy change
------------------------

IDA plugin will not any longer try to locate IDA with locate or find.
Both proved to be slow and error prone. Instead, it will look for it
in the PATH, and if nothing found, will refuse to provide a service.

It is still possible to provide an explicit path to IDA, but the correct
way is to determine the path to ida, at the configuration time.

This will close #404.

New source infrastructure
-------------------------

Previously we had only three types of source, encoded with GADT. This PR
allows to specify sources of arbitrary types, due to a switch from
inductive sources of type `unit -> 'a` to co-inductive of type `'a ->
unit`. In particular a data provider, can capture arguments of any type
in the closure, and will communicate the result, to a data consumer
using a callback mechanism, that is wrapped into a coinductive streams
and futures.

New plugins policy
------------------

Now a plugin registers itself as soon as possible (instead of waiting
for the inputs), so a user have a real information on what is
available. But if a plugin is chosen explicitly by a user, for example,
with an option `--rooter=ida`, and the plugin was unable to provide the
data, then it should fail with a diagnostic message, instead of trying
to provide an empty input. This gives a much better feedback and user
experience.

Byteweight error handling
-------------------------

Byteweight will now be much more careful with error handling. This
closes #342.