Fix PE32 support and few bug fixes. (#418)
There were few bugs in the LLVM, all in the upstream. This
PR fixes them on our side.

* Address of the segment was incorrect
* Address of symbol was incorrect
* Size of symbol was incorrect

The first issues was due to the use of the relative virtual address,
instead of virtual. They are fixed by adding the image base address.

The latter issue was quite funny. The algorithm that calculated address
in llvm was = "subtract symbol address from section size" (sic). As an
excuse the author wrote a comment "FIXME: Return the correct size."
Indeed, it was quite hard to get the correct size, as the correct
algorithm is quadratic.

But now we can discover symbols from PE32, even without usage of
external tools like IDA or objdump.

Two other bugs were:

1. objdump plugin didn't work for 32-bit binaries, as addresses was
always 64 bit.

2. internal rooters and symbolizers didn't work due to a logic error in
the DNA of the author.