detect binutils installation at configuration time (#422)
Alike as IDA, objdump now expects a path to objdump to be passed
at configuration time. A corresponding `conf-binutils` package will
be added to opam and the dependency between them will be established.

As a consequence, `bap-objdump` package will not be installable, if
there is no objdump program, moreover, a corresponding external
dependency will be added and thus installed via depext.