added conf-{ida,objdump} dependency to opam/opam
Since someone may use this file for obtaining bap, I've added an
integration with `conf-ida` and `conf-objdump` packages. These
packages are only available in our opam-repository, so don't forget to
add it. To get working ida or objdump plugins on your box, make sure
that these packages are installed. To get proper objdump on your linux
or mac box, invoke `opam depext conf-objdump`.

Note -- after this PR is merged, ida and objdump will become unavailable
if conf-ida and, correspondingly, conf-objdump are not installed.