add ida powered file loader
With the new loader it is possible to load all files, that are
supported by ida, including:

- kernel modules (closes #345)
- object files (closes #402)

Note - the default loader is still llvm, to switch to the ida loader,
pass --loader=ida to bap.

This PR also fixes a bug in the project construction function, we left
out the code segment in the final project.

Also, this PR adds support for writing ida scripts in IDC, and rewrites
symbols command in IDC.