fix microx backtracking (#455)
The backtracking point was registered on the entry to a block,
as a result, when we were backtracking all updates to the context
made in the last block were ignored.

A better way would be to store the context just before the first
jump. To simplify the implementation we will store context actually on
every jump (since usually there one or two jumps on a block, this
wouldn't have a big impact). Also, for performance consideration, an
associative list was substituted with a real Tid.Map.