Allow a frontend to link with internal libraries.
Previously it was disallowed, see the big fat warning in the
`oasis/fronted` (removed by this PR).

This PR will now track both internal and external libraries. We leverage
the predicate mechanism of ocamlfind, and for each internal library
`LIB` thatis linked with a host program, we record a predicated
`used_LIB`. Later, we can extract the names of linked libraries and
record them in the registry of linked libraries.

This PR also revives the byteweight frontend, that was broken due to a
dependency on the internal library byteweight (the problem manifested
itself only if it was installed from the source tree, it worked ok from
the opam-repository).