better IDA debugging (#480)
Now it is possible to debug IDA with `BAP_IDA_DEBUG` environmental
variable. If it is greater than `1`, then an autonomous mode will be
disabled, and in case of error an error dialogue will pop up. If it
is greater than `2`, then the library will not clean after itself,
leaving databases,scripts and all the intermediate stuff.