a more intensive target lookup in llvm backend
This patch adds a support for systemz architecture. For some reason,
if passed without name, a target is found, but RegInfo structure is
not. On the other hand, all other architectures, except, mips, can't be
found at all if an architecture is passed explicitly. So, we try both,
first search using official method with both arch and triple specified.
If it fails, then specify only triple.