Better target lookup in llvm backend (#530)
* better target lookup in LLVM backend

The new lookup stategy follows closely llvm-mc implementation. It is
able to find all existing architectures, and systemz, that we weren't
able to use before.

* a more intensive target lookup in llvm backend

This patch adds a support for systemz architecture. For some reason,
if passed without name, a target is found, but RegInfo structure is
not. On the other hand, all other architectures, except, mips, can't be
found at all if an architecture is passed explicitly. So, we try both,
first search using official method with both arch and triple specified.
If it fails, then specify only triple.

* updated testsuite