Merge prefixes in llvm backend (#532)
* disable prefixes in LLVM

In the disassembly step if a prefix is detected[1] then we ignore it and
recurse into step function again. When we return we add a memory region,
occupied by a prefix to the instruction. Such scheme allows arbitrary
nesting of prefixes.

[1]: Since there is no official way (i.e., a predicate) for identifying
that an instruction is a prefix, we rely on a convention that all prefix
instructions has a `_PREFIX` suffix in their names. For performance we
build a set of instruction opcodes, that has such names.

So far there is only one thing TBD - to should fail with invalid
instruction if a prefix is not followed by a valid instruction.

* uint64_t for addr and more robust step function

The step function now checks for errors more carefully and doesn't trust
llvm at all. Now llvm backend will not emit prefixes at all (an empty
prefix or a prefix with a prefix is an invalid instruction). As well, a
prefix to invalid instruction is also invalid.

* update testsuite

* Fixed prefix behavior is special cases.