publish create/close functions for basic disasm
Previously a basic disassembler could be only accessed by `with_disasm`
function, that is safe but not very versatile. There is at least one use
case, where the CPS style is not enough, as we may need to have a pool
of disassemblers. For example in bap-server.

This PR adds this functions to the module interface, however they are
discouraged, and `with_disasm` function should be considered whenever
it is possible.