support code for the new python interface
The new python interface uses popen to run bap and dump project data in
the ADT format, that is readable from python. This simple approach
allows to get a programmatic access from python to all the data, that
was discovered during the analysis. The following example, demostrates
the basics:

```python
>>> import bap
>>> proj = bap.run('/bin/true', ['--symbolizer=ida'])
>>> text = proj.sections['.text']
>>> main = proj.program.subs.find('main')
>>> entry = main.blks[0]
>>> next = main.blks.find(entry.jmps[0].target.arg)
```

See the corresponding repository for more documentation. It is
recommended to use ipython to explore the bindings, later we will create
a sphinx documentation.

Changes:

1. added ADT printers for the project data structure
2. for consistency dashified all attributes that contained underscores
3. removed subroutine-addr, insn-addr, block attributes, substituted
   them all with the `address` attribute (as there is no need to
   distinguish between instruction or subroutine or block address).
4. removed few unecessary attributes, like `target-name` or
   `subroutine-name`.

Note: bap-python repository will be updated as soon, as this PR is
merged.