generalized interpeters with a monad type
This commit will not touch the API, as bap.mli is remained
unchanged. The commit will serve as a testimony that the API is
unchanged, as all changes satisfy unchanged bap.cmi.

Ideally, we should make a CI test, that will check, that new cmis are
subsests of the old one (or that new implementations satisfy old cmis).