invoke Lisp_machine properly from the interpreter
Now, the interpreter is responsible for binding arguments and evaluating
the return value. It still looks like, that we're drawing an incorrect
abstraction boundary, as the interpreter is now taking to much
responsibilities, that it will not be able to fulfill. For example,
given more complex calls with hidden arguments we will need to have a
complex interaction between function declarations and ABI. For example,
if a function returns a structural object, or it is a C++ member
function, then it will have a hidden parameter, that should be declared
from the lisp side, e.g.,

     (defun make-point (x y)
        (declare (hidden self))
        (struct-write point_t self x x)
        (struct-write point_t self y y))

The `struct-write` should be a builtin function, that is implemented
in OCaml, and calls to existing ABI/API stuff.