propagate backtraces from passes
Allow users to see a backtrace when an exception is thrown
from a pass.

The backtraces are printed for any exceptions, including Failure and
Invalid_arg (i.e., special handlers for these exceptions are disabled).

The problem
===========

The pass runner is capturing exceptions that are thrown by a pass, and
rethows it wrapped into the `Pass.Failed` exception. Since the newly
thrown exception is physically (and structurally) different from the
caught exception the backtraces is erased and is not propagated any
further. As a result, the printed backtrace shows only the path from the
pass runner to the entry point of the frontend. That is always the same
and does't really bear any information.

Obstacles
=========

Since the BAP API is fronzen, we can't change the `Pass.error` type, or
the `Failed` exception.

Solution
========

The backtraces is extraced by the pass runner at the place where it is
caught and translated into a string, then it is attached to the
`Exn.Reraised` exception, and reraised. There is a room for improvement,
however. First of all, we're sort of abusing the `Exn.Reraised`
exception. However, the abuse is local, and other solutions, would not
be anymore safe or robust (other then an api-breaking change, that will
affect the representation of the exception or the error type). Second,
the propagated backtraces still has few slots, that are specific to the
pass runner and do not bear any interesting information. It should be
nice, to trime them away, to get a clean trace, that starts from the
user plugin.