matures beagle
This is a big commit, reflecting several days of beagle training.

First of all, I've trained him to understand few new commands:

--beagle-print-strings - will print all found static strings;
--beagle-print-words - will print all the generated words;
--beagle-print-chars - print the detected char sequences;
--beagle-ignore-strings - will ignore found static strings;
--beagle-no-words - will not try to build words from a dictionary;
--beagle-alphabet - will force beagle to use different alphabet;
--beagle-words - will add words to beagle's dictionary.

The default use case is still the same - run beagle via `--pass=beagle`
and annotate program tree with references to strings.

The print options are mostly for the debugging and standalone using.

Speaking of the static strings, i.e., if we will compare it with the
`strings` utility, currently beagle outputs slightly less strings than
`strings`. This is not a bug, or limitation. As by design Beagle finds
only strings that are null-terminated, where `strings` will find all
strings even if the last character is an arbitrary non-printabl byte. To
my experiments, this behavior only adds more trash. However, probably,
there was a good reason why `strings` authors decided to use any
strings. For example, to work with pascal binaries, where strings are
not required to have a null character at the end. Given this
consideration, we will probably resign from our allegiency to the
null-terminator in the future.