anticipates a non-trivial destination in bapbundle
The desintation may not exist, in this case a default behavior of the cp
utility would be to treat the destination as a file. So we need to
create a directory first, if it doesn't exist. It is also possible, that
the destination is not a folder, and in that case we shouldn't overwrite
it with the plugin (as it is usually not what a user wants), but instead
we will ask for a directory name.