binary deployable bap (#623)
* make bap more standalone

wip

* switches to native version of postinstall tool

as the byte version doesn't work under sudo.

* adds plugindir variable

Adds a new specific configuration variable that designates a location of
plugins, instead of $libdir/bap as it was before.

* anticipates a non-trivial destination in bapbundle

The desintation may not exist, in this case a default behavior of the cp
utility would be to treat the destination as a file. So we need to
create a directory first, if it doesn't exist. It is also possible, that
the destination is not a folder, and in that case we shouldn't overwrite
it with the plugin (as it is usually not what a user wants), but instead
we will ask for a directory name.

* cleans debug print statements

* adds a script for building deb packages

The script is extremely crude, and doesn't really follow debian or
ubunty policies, but it is better to have something.

The produced debian packages have very few dependencies, basically,
libz, libgmp, and libstdc++. No OCaml or LLVM dependenices (as they are
all statically linked).

The script will build three packages:

* bap.deb - bap frontend and plugins;
* libbap.deb - bap C bindings shared library;
* libbap-dev.deb - development files for bap C bindings.

They can be installed with:

dpkg -i bap.deb
dpkg -i libbap.deb
dpkg -i libbap-dev.dep

The script will create a new opam switch and make fresh clones of bap
and bap-bindings repositories.