uses all available symbolizers by default
We used to have this behavior before, but at some point we started to
default on the internal symbolizer (that is dump). I can't recall the
reason, but I hope that there was one. Nowadays, I don't see any
sufficient reasons for not using all the provided information. Probably,
originally we used different merging strategy - now the strategy will
take the first name, that is known to the symbolizer, and will fall back
to a `sub_XXXX` if and only if the name is not known.

As a result, now it is not needed to specify explicitly
`--symbolizer=ida` to get good naming, if IDA is installed, it will be
used automatically.

There is, however, some caveat - if two symbolizers are providing
different names to the same subroutine, then, depending on a order of
installation (basically random), subroutines will be named
differently. Probably, we should invent some mergin strategy that will
collect all naming suggestions and deal with the ambigious names.