uses beta-reduction for evaluating let-expressions
The previous implementation used host's stack to store previous values
of a let-bound variable. However, such behavior, leads to an unnecessary
side effects, like an access to an undefined variable. Even worse, after
an evaluation of a let-expression the environment will be poisoned by a
new non-existent binding.

The new implementation follows closely (if not to say literally) the
semanitcs of the BIL language. The let expression is first beta-reduced,
i.e., all bindings are basically inlined. As a result, the let-bindings
are now more of a syntactic sugar. We probably should publish this
function, so that it can be used in other analyses, that don't want to
reason about let-expressions.

The beta-reduction is applied in a clever way that doesn't require an
alpha-conversion. Basically, the bindings are pushed into a
stack (finite map), and applied in the order (exactly as it is described
in the operational semantics).