adds heuristics for identification of `main'
The heuristic is simple, the main function is the function that is
passed to the `__libc_start_main` function. If the latter name exists,
then we can definitely find the `main` function. Of course, this
heuristic relies on a presence of `__lic_start_main` that is probably
specific to the gcc/glibc, but we will try to find corresponding
functions for other ABI.