links Lisp machine via the linker interface
The Interpreter is now independent of the Lisp machine and uses only the
linker interface to call subroutines. The Lisp machine is added as a
regular machine component via the linker interface. The linked code is
now parametrized by an interpreter instance.

Also the Lisp machine tries to evaluated as much as possible during the
linking phases, so that the errors are not delayed until the runtime.