primus now emulatates our test program
In fact it is even able to emulate an extended program, that I will
commit to the testsuite soon. The program uses malloc, free, memmove,
strlen, strcpy, strchr, and putchar to copy and parse user input, and
print it out.

Performance-wise, the testing programs runs twice as faster than qemu.