adds more docs and few changes to the interface
We now more or less distinguishing between an error that denotes that a
document is not well-formed and a system error. They should not be
merged. Currently we're raising an exception in case of the latter, this
however brings us into a world of doomed who raises an exception from a
function that returns a result. Probably, in the future we will switch
from the Or_error.t, to some error type, that distinguishes between
different kind of errors.

Note, so far the documentation describe the desired behavior, not the
actual one. We still have some work to do there (actually type-checking
is not implemented).

This commit also adds a [merge] function, that takes two documents and
merges them.