adds a self-join operation
This commit makes it possible to select the same attribute multiple
times (i.e., to make a self join). It also provides a new kind of
variables - the positional variables. A string-like subscript is used to
create such variable, e.g., `fld.[n]` will bind a field [fld] of the
[n]'th attribute in a selection, e.g., to find all different students
with the same name:

```
      {[select (from students $ students)
          ~join:[[field name]]
          ~where:(id.[0] <> id.[1])]}
```

This commit also fixes a bug, that was triggered when an empty attribute
was selected as a part of a bigger selection.