switches to a normalized form to fasten comparison
this commit adds an equality operator that is applied to fields without
parsing them. This removes a constant factor to less than a microsecond
per entry. The algorithm is still exponential and no query optimization
(e.g., reordering and filtering before product) is done. This the field
of the future improvements.