passes both virtual and physical sizes to loader
This commit adds a small workaround for the legacy loader, with which we
can pass both the virtual and the physical size of a segment. The idea,
is that for each segment in a file, we create the same named section,
that has the same base address, but the size field is filled with the
virtual size of a segment. The solution is backward compatible with the
legacy backend behavior, hence if the virtual section is not provided,
the physical size will be used.

Implementation
--------------

1. On the backend side, a vsize field was added to the segment data
structure. The old trick with negative offsets is removed, instead both
sizes are stored in the segment data structures. After an image is
created, for each segment we add a section that has the same name and
base address, but uses the vsize field as the size.

2. In the Image module, when a mapped region is created we search for
the section that has the same name and base address as the segment, that
we are constructing. If such section is found, then we use its size as a
segment size, otherwise we fallback to the size that is provided by the
segment.

P.S. All tests are passing! (At least on a machine with IDA).