revises containers operations in Monads
Apparently, what I was missing is that a chain of bindings of a
non-eager monad consumes the stack memory proportionaly to the length of
the chain. It is not a big issue for a hand-written chains, but in a case
of chains built from collections of elements, it will eventually (when a
monad is run) raise a Stack_overflow.

Another problem, that this commit fixes, was the usage of the
`with_return` construct inside of a monad. Again, it worked fine with
eager monads, but fails dramatically.

To solve the first problem, we switched to CPS implementation of fold
function lifter into a monad. This is, however, insufficient for the
second problem, as we need to have a mechanism, that will allow us to
escape from a collection processing loop earlier. (We used exceptions
for this, but exceptions don't play well with monads, that establish
their own contol flow). To address this problem, we changed the
[Collection.Base] interface. In fact we introduced two new interfaces,
the [Collection.Eager] deduces all collection operations just from an
ordinary [fold] function. The [Collection.Delay] provides a [foldk]
function, that applies a user function with a continuation, that can be
used to continue looping if necessary, or to escape from a loop if not.
We decided to provide to functors, since it is not always possible to
provide [foldk] for some containers, especially to those, that are
abstract, and that do not provide anything similar.

This commit also removes the `rev_` family functions from the Container
interface, mostly because the new map, filter, and filter_map
implementation uses the fold_right function, so there is no big
efficiancy difference between map and rev_map. (Though, I maybe wrong).