enhances invoke-procedure, makes symbols visible
The type is no longer strictly needed to invoke a procedure via this
primitive. Instead of looking at the C type, we are using the BIR type,
and the caller must provide the arguments based on the subroutine IR.

This commit also implements the propagation of a return value from the
invoke-procedure, for the (unlikely) case, when it actually returns to
the caller.

Another enhancment is the new mechanism that makes possible to reference
to functions (and other symbols) by name, instead of finding its
address. Bindings from all terms names to their addresses are added to
the initial context.