extracts semantics from BIL interpreters
This commit extracts an abstract (virtual in OCaml parlance)
implementation of expression interpreter. The interpreter, in named
`eval` and is a base class for the `expi` interpreter. The interpreter
is implemented as a virtual class, requiring the users to implement two
virtual interfaces: `eff` and `domain`.

The implementation of the `['a,'r,'s] eff` interface handles all the
side effects, like memory and environment operations. The ['r] type
variable, denotes the domain of values. The ['s] type variable denotes
the type of memory storage.

The implementation of the `['a,'r,'s] domain` must implement methods
that map the domain of computation `'r` to bitvectors and storages.

The `eval` interpreter doesn't depend on the concrete representation of
value and storage, such as Bil.result and Bil.storage, and thus is
suitable as a new base implementation for the primus interpreter, that
previously was based on a Biri intererpreter, that operates on an
improper domain (Mem and Bot variants are not used), and doesn't use
storage mechanism at all (the storage value is ignored, and data are
stored in internal state, bypassing the `storage` value). The `eval`
interpreter doesn't depend on a context, that reliefs Primus from a
burden of passing this attavims across its components.

Basically, if we will move Primus on the new eval interpreter, the
implementation of Primus should become much cleaner and
smaller (especially type signatures).

In the next step we are going to get rid of the dependency on the Biri
from Primus, as Biri is not suitable for the non-deterministic
computation.